Detect the tip of the dart
The intersection between the axis of the dart and the dart board level gets
calculated with the all the information that is collected for the processed
image.

Refactoring: added a data structure to hold all information of a captured image.

Refactoring: Extracted all functionality to display the result in a seperate file